move most of expand_cfg_attr_item to a method on `Attribute`
